Коммит №6 Убран пустой модуль, методы с логикой перемещены в helpers. Тесты успешно проходят.
